--- a/x-pack/plugins/apm/scripts/apm_load_testing/.gitignore
+++ b/x-pack/plugins/apm/scripts/apm_load_testing/.gitignore
@@ -1,4 +1,4 @@
-./dist
+dist
 node_modules
 screenshots/
 .env