--- a/x-pack/plugins/apm/scripts/apm_load_testing/package.json
+++ b/x-pack/plugins/apm/scripts/apm_load_testing/package.json
@@ -1,0 +1,15 @@
+{
+  "scripts": {
+    "start": "ts-node --transpile-only src/index.ts"
+  },
+  "dependencies": {
+    "@types/lodash": "^4.14.180",
+    "axios": "^0.26.1",
+    "dotenv": "^16.0.0",
+    "lodash": "^4.17.21",
+    "p-retry": "^4.6.1",
+    "puppeteer": "^13.5.1",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.2"
+  }
+}