--- a/x-pack/plugins/apm/scripts/apm_load_testing/src/index.ts
+++ b/x-pack/plugins/apm/scripts/apm_load_testing/src/index.ts
@@ -128,7 +128,7 @@ function getAverageRequestTimings() {
     ([name, obj]) => {
       return {
         name,
-        avg: meanBy(obj, (o) => o.endTime),
+        avgEndTime: meanBy(obj, (o) => o.endTime),
       };
     }
   );