This pull request introduces a new tool for doing ad-hoc benchmarks of APM endspoints in Kibana. The intention is to make performance improvements and regression tangible by comparing the performance results from before and after the change.

The tool is implemented specifically for measuring the performance improvements made to the Inventory page, but should be generally useable.

*Disclaimer: the tool is highly POC. It's very simple, and depending on the scenarios that are being tested, and environmental factors results will vary. We should therefore validate the numbers before acting on them.*

## Methodology

- The test opens the inventory page and waits for all apm api calls to complete. Tests are run automatically 50 times, and returns the overall average "end time" for each api request. `endtime = start.time + duration`. End time is useful because it's an indicator of how early the user sees data. It no good that a request has a very short duration if it's queued up for 10 seconds. 
- Load tests were performed against a local Kibana cluster, with data from edge cluster via ccs. Testing against an adequately sized cloud cluster may provide more accurate results.
- Elasticsearch cache is cleared between every run to simulate worst case. 
- The script was executed against latest commit on `main` (067f4c2edeccdd4a917a86d0e0f6ac8c39c9e786) which contains the perf improvements, and compared against an older commit that didn't contain the perf improvements (c8a206545a99bcd96a754ecfcf9380e2a80ffdd2).
- For this test we are only considering the endpoints `/internal/apm/services`, `/apm/services/detailed_statistics` and `/apm/sorted_and_filtered_services`. 

## Hypothesis

The perf improvements adds a new endpoint `/internal/apm/sorted_and_filtered_services` which uses the terms enum api and should be much faster than the terms agg uses previously to fetch service names.
The expectation is that the table will be populated with service names very quickly (hopefully in less than 2 seconds). The remaining data (metrics and sparklines), will then be loaded lazily. Whether their load times increase or decrease slightly is not a concern in this phase.

## Results

### Before perf improvements

```js
[
  {
    // load service names and metrics
    "name": "http://localhost:5601/internal/apm/services",
    "avgEndTime": 10917
  },
  {
    // load sparklines  
    "name": "http://localhost:5601/internal/apm/services/detailed_statistics",
    "avgEndTime": 15213
  }
]
```

### After perf improvements

```js
[
  {
    // load service names
    "name": "http://localhost:5601/internal/apm/sorted_and_filtered_services",
    "avgEndTime": 9862
  },
  {
    // load metrics
    "name": "http://localhost:5601/internal/apm/services",
    "avgEndTime": 11744
  },
  {
    // load sparklines
    "name": "http://localhost:5601/internal/apm/services/detailed_statistics",
    "avgEndTime": 16135
  }
]
```

## Analysis

### Service names
Before: 10917ms
After: 9862ms

The new API (`/internal/apm/sorted_and_filtered_services`) returns service names roughly 10% earlier than the existing one (`/internal/apm/services`). 
We had hoped for a much bigger improvement. The lack of improvements might be related to running on an under-ressources local cluster. We should try repeating the test on an adequately sized cluster

**Result: 10% improvement**

### Metrics (latency, throughput and failure rate values)
Before: 10917ms
After: 11744ms

We haven't done anything to improve this, so is as expected

**Result: 7% regression** 

### Sparklines
Before: 15213ms
After: 16135ms

We haven't done anything to improve this, so is as expected.

**Result: 6% regression**

## Out of scope

 - Tests currently don't take browser rendering into account since the majority of time is spent in Elasticsearch. This could  be useful to explore for other scenarios in the future.
 - I initially tried to use the RUM agent for capturing performance metrics but custom instrumentation was necessary to make that work. We might still go in that direction in the future.